WooCommerce 2.5 + refactoring mailtemplates
